refactor(r4a): move guestaudit to internal/security (was services/shell)
R4a reorg. The botanical "shell" (security posture) area had only one real,
api-used package left after the facade deletion (#1441): guestaudit. Move it to
a descriptively-named security domain:

  internal/services/shell/guestaudit  ->  internal/security/guestaudit

Pure move + import-path rewrite (1 importer: handlers_guest_audit.go). The
internal/services umbrella is now reduced to just discovery (HOT — promotion to
internal/discovery deferred to Phase 6).

Behavior gate: golden HTTP suite + guest-audit route unchanged. go build ./... +
GOOS=windows green; go test ./... green; golangci-lint 0.
