+++ internal/security/guestaudit/ping_windows.go
